Add and delete tasks to projects including inbox
Tasks can now be added and deleted in a project (backend). If the task
is added to a project other than 'Inbox' then the task is added to Inbox
as well. If a task is deleted in one project and it belongs to other
projects, it will be deleted in all other projects. In a future commit,
tasks will be added to and deleted from 'Today' and 'This Week'
depending on the task's due date.